+++ addon/components/fd-ember-jstree.js
@@ -65,6 +65,15 @@ export default Component.extend({
   })),
 
   /**
+    Multi select.
+
+    @property multiple
+    @type Bool
+    @default false
+  */
+  multiple: false,
+
+  /**
     Data for search tree nodes.
 
     @property searchTerm